fixed compiler warnings

+++ amqpcpp/decimalfield.h
@@ -125,7 +125,7 @@ public:
      */
     operator float() const
     {
-        return _number / pow(10, _places);
+        return static_cast<float>(_number / pow(10, _places));
     }
 
     /**
